upgrade cart functionality

+++ app/public/js/main.js
@@ -2,7 +2,7 @@ $(document).ready(function() {
     var sum = 0;
     $('.price').each(function() {
         var price = $(this);
-        sum += parseInt(price.html());
+        sum += parseFloat(price.html());
     });
 
     $('#total_price').html('Total: ' + sum);
